Move obsolete plugins to obsolete directory
Move obsolete plugins to obsolete directory
